feat: Add user management components including JobTitleChangeConfirm, CustomPagination, PermissionCheckbox, and UserTable
